requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27
pydantic-settings>=2.2
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
alembic>=1.13
httpx>=0.27
tenacity>=8.2
python-multipart>=0.0.9
orjson>=3.9
email-validator>=2.1
redis>=5.0
hiredis>=2.3

# Dev dependencies
ruff>=0.4
pytest>=8.2
pytest-asyncio>=0.23
anyio>=4.0
